runner: move thread processing infra to libtcmu
